Support lexicographical sorting
Allow to return the largest tag after lexicographical sorting. This can
be usefule in returning largest semver tag. Main use case can be in the
projects where multiple versions are supported. For example, `v2.38.6`
is the latest stable version but a recent security patch was pushed in
version `v1.15.8`. Thus, it is now possible to sort in lexicographical
order and get the highest semver version (`v2.38.6`).

Signed-off-by: rustyclock <rustyclock@protonmail.com>